\name{Perform PPR, PCA, or LDA on a dataframe}
\alias{DimReduction}
\title{
  DimReduction
}
\description{
  Performs projection pursuit regression (using ppr from stats), linear discriminant analysis (using lda from MASS), and principal component analysis (using prcomp from stats) on a dataframe after including or excluding certain columns. The PP1, PP2, PC1, PC2, LD1, and LD2 directions are then dotted with each row to produce PP1, PP2, PC1, PC2, LD1, and LD2 columns in the new dataframe.
}
