\name{Density surface}
\alias{TPS_distribution}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  TPS_distribution
}
\description{
  Create a density surface given inputs to binCounts using TPS. This is the 2-variable equivalent of a density curve, which goes along with a histogram.
}
