\name{Generate a fitness landscape}
\alias{TPS_landscape}
\title{TPS_landscape}
\description{
  This creates a fitness landscape using thin plate spline (TPS) interpolation using the fields package. Various "output" arguments are wireframe, contour, matrix, model, and plotly (which generates a plot the user can interact with).
}
